# Conflicts:
#	Robust.Shared/GameObjects/Systems/SharedUserInterfaceSystem.cs

+++ Robust.Shared/Timing/IGameTiming.cs
@@ -86,7 +86,7 @@ namespace Robust.Shared.Timing
         /// <summary>
         ///     The target ticks/second of the simulation.
         /// </summary>
-        byte TickRate { get; set; }
+        ushort TickRate { get; set; }
 
         /// <summary>
         /// The baseline time value that CurTime is calculated relatively to.
@@ -167,7 +167,7 @@ namespace Robust.Shared.Timing
         void ResetSimTime();
         void ResetSimTime((TimeSpan, GameTick) timeBase);
 
-        void SetTickRateAt( byte tickRate, GameTick atTick);
+        void SetTickRateAt(ushort tickRate, GameTick atTick);
 
         TimeSpan RealLocalToServer(TimeSpan local);
         TimeSpan RealServerToLocal(TimeSpan server);
